refactor: adjust name trusted to healthy
