+++ frontend/src/components/HealthyRepoCard.svelte
